Redesign the OLED user interface, implement USB programming
Due to testing with a hand soldered prototype, the original design of
using a five-way navigation tactile switch for the display navigation
was removed due to the button's fragility. A push button can be
registered as a push, down, and right command. The display will be
navigated using the PS3 controller, as that was already accessible.
USB-C periphials were added for more universal firmware updates, instead
of requiring the use of the ESP-Prog.